requirements.txt
bitstruct==8.10.0
certifi==2020.6.20
chardet==3.0.4
coloredlogs==14.0
docopt==0.6.2
erl-terms==0.1.5
guietta==0.3.3
humanfriendly==8.2
idna==2.10
paho-mqtt==1.5.0
parsimonious==0.7.0
Pygments==2.6.1
PySide2==5.15.0
requests==2.24.0
shiboken2==5.15.0
six==1.15.0
urllib3==1.25.10

recordclass~=0.14.3
